Break out app/inc.hpp into component header & src files
