+++ firmware/projects/LVController/app/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(main PRIVATE
+    subsystem.cc
+    dcdc.cc
+    fan.cc
+)
